experiments/osrm/pyproject.toml
[project]
name = "osrm-preprocessing"
version = "0.1.0"
description = "OSRM data preprocessing pipeline"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "boto3>=1.34.0",
    "pytest>=7.4.0",
    "loguru>=0.7.3",
]

[project.scripts]
download-osm = "src.core.download_osm:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
